chore: extract utility to read ProContext
